ai-service/requirements.txt
flask==3.0.2
flask-cors==4.0.0
transformers==4.38.1
torch==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
